services/segment_api/pyproject.toml
[tool.poetry]
name = "segment_api"
version = "0.1.0"
description = ""
authors = ["kyle <kyle@jataware.com>","graham <graham@jataware.com>"]
readme = "README.md"
packages = [{ include = "segment_api" }, { include = "segment_api/**/*.py" }]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.23.2"
fastapi = "^0.109.0"
pydantic-settings = "^2.0.3"
rasterio = { version = "1.3.8", extras = ["s3"] }
numpy = "1.26.2"
python-dotenv = "1.0.0"
cdr_schemas = { git = "https://github.com/DARPA-CRITICALMAAS/cdr_schemas.git", tag = "v0.4.12" }
boto3 = "^1.28.52"
python-multipart = "^0.0.6"
pillow = "^10.1.0"
cachetools = "^5.3.2"
httpx = "^0.26.0"
redis = "^5.0.5"
opencv-python = "^4.10.0.84"
opencv-python-headless = "^4.10.0.84"
transformers = "^4.41.2"
torch = [
  { platform = "darwin", version = "^2.4.0", source = "pypi" },
  { platform = "linux", version = "^2.4.0", source = "torch-cpu" },
]
tifffile = "^2024.7.24"
imagecodecs = "^2024.6.1"

[tool.poetry.scripts]
dev = "dev.run:main"
embeds = "dev.embeds:main"
format = "dev.script:format"
lint = "dev.script:lint"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
flake8-print = "^5.0.0"
autoflake = "^2.2.0"
uvicorn = "^0.23.2"
pytest = "^8.2.2"

[[tool.poetry.source]]
name = "torch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
